Refactor : Move auth files
