+++ src/hooks/useSetUserAndNavigate.ts
@@ -1,7 +1,7 @@
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { setUsername } from '../features/user/userSlice';
-import { LoginRepRes } from '../repository/type';
+import { LoginRepRes } from '../repository/auth/type';
 
 export const useSetUserAndNavigate = () => {
   const dispatch = useDispatch();
