+++ src/repository/auth/auth.ts
@@ -1,5 +1,5 @@
-import { config } from '../config.js';
-import HttpClient, { ErrorResponse } from '../network/http';
+import { config } from '../../config.js';
+import HttpClient, { ErrorResponse } from '../../network/http.js';
 import {
   LoginRepRes,
   LoginResReq,
