+++ src/repository/auth/type.ts
@@ -1,0 +1,8 @@
+export type LoginResReq = {
+  username: string;
+  password: string;
+};
+export type LoginRepRes = { user: { username: string } };
+
+export type SignupRepReq = LoginResReq;
+export type SignupRepRes = LoginRepRes;
